+++ src/model2/index.ts
@@ -2,6 +2,7 @@ export * from "./account";
 export * from "./account_thresholds";
 export * from "./account_flags";
 export * from "./account_values";
+export * from "./account_subscription_payload";
 export * from "./asset";
 export * from "./asset_input";
 export * from "./data_entry";
@@ -10,12 +11,15 @@ export * from "./data_entry_values";
 export * from "./ledger";
 export * from "./ledger_header";
 export * from "./mutation_type";
+export * from "./native_trust_line_subscription_payload";
 export * from "./offer";
 export * from "./offer_values";
+export * from "./offer_subscription_payload";
 export * from "./transaction";
 export * from "./transaction_with_xdr";
 export * from "./trust_line";
 export * from "./trust_line_values";
+export * from "./trust_line_subscription_payload";
 export * from "./signer";
 
 export * from "./operation";
